ebpf: fix all headers without changes
Current eBPF source headers have several problems:

- some of them use local headers "", some the search path <>
- clangd wasn't able to index headers correctly

Follow: #2359 for TODOs

NO CHANGES BUT POSITIONAL AND INCLUDE FIXES.
